+++ app/src/main/java/com/dogandpigs/fitnote/presentation/lesson/loadlesson/LoadLessonState.kt
@@ -4,7 +4,7 @@ import com.dogandpigs.fitnote.data.source.remote.model.Info
 import com.dogandpigs.fitnote.data.source.remote.model.LessonInfo
 
 data class LoadLessonState(
-    val selectedUserId: Int = 0,
+    val selectedUserId: Long = 0L,
     val memberList: List<Info> = listOf(),
     val exerciseList: List<LessonInfo> = listOf()
 )
